ecore-evas: call application-set focus functions if no engine functions exist
in the case where an engine has no real concept of focus (eg. drm), no engine
functions will be implemented, resulting in calls to focus_set having no effect.
this leads to elm/applications being unable to receive the callbacks they expect
when calls to the overall api are made, resulting in focus being broken

probably this should also be done for the rest of the api functions too

@fix

Change-Id: Ie3dfa6518d9b7375a65a9ff1c963715ecaff47ab
Signed-off-by: Mykyta Biliavskyi <m.biliavskyi@samsung.com>