wayland_shm: Add swap mode fallback when buffer age changes
This copies the behaviour in opengl_x11 engine, where the buffer
age needs to be continuously the same to be taken into account.
If the age varies, then we fallback to a full redraw.

Apparently this fixes issues on actual devices. I tested this
patch in weston (I didn't have issues before and buffer age is 1).

This commit is cherry-pick of the upstream commit.
https://git.enlightenment.org/core/efl.git/commit/?id=62ca4486ea30fd63a5589ea7af373a5cb9982031

TIZEN_ONLY tag has been removed

Change-Id: If479243d5f8bce5bd41142e833e04602e6152326
Signed-off-by: Wonki Kim <wonki_.kim@samsung.com>