ecore_wl: Add values for indicating a position configured by server.
These values will be updated when a position of client is changed by server,
and it will be used to compare with arguments of ecore_wl_window_position_set().
The reason why this is needed is to avoid set a undesirable position
if client try to set position again before receiving changed message from
server for first one.

Change-Id: Iab25e1830250005c33dc72459c3d2ad152a5b4f1