Add -Wno-strict-aliasing
-Wstrict-aliasing
This option is only active when -fstrict-aliasing is active. It warns
about code that might break the strict aliasing rules that the compiler
is using for optimization. The warning does not catch all cases, but
does attempt to catch the more common pitfalls. It is included in -Wall.
It is equivalent to -Wstrict-aliasing=3

Signed-off-by: Dongkyun, Son <dongkyun.s@samsung.com>