Edje entry: Do not update anchors outside viewport.
Summary:
Do not update anchors outside viewport.

Problem is that when entry contains large number of anchors,
scrolling and launch performance is degraded due to lots of objects
being created and maintained, so to enhance the performance of
entry with anchors, anchors is not updated outside the viewport.

Test Plan:

    Should have large number of anchors in entry
    Observe the launch and scrolling performance.

Reviewers: tasn, raster, herdsman

Subscribers: thiepha, raster, cedric, jpeg

Tags: efl

Differential Revision: https://phab.enlightenment.org/D3543

Change-Id: I49cf544f3caad9a0e3cf5a9157906a7347eb5af1
Signed-off-by: SubodhKumar <s7158.kumar@samsung.com>