[EvasGL] Fix to 3D memory leak.
If evas_gl_make_current is called with NULL,
and evgl_surface_destroy and evgl_context_destroy is called,
3D memory leak occurs.
Because rsc->current_ctx is NULL in evas_gl_make_current
and texture and framebuffer can not be free in destroy funnctions.
To fix that, one value is added to EVGL_resource.
This value is kept with current_ctx, when evas_gl_make_current is called with NULL.

Change-Id: Iabb7685d0c5e932b0e08e716225bf15f58696d0b