eina_file_win32: close handler in case of error
Func _eina_file_win32_first_file try to find the first file in directory
but if any file not found the file handler stay open, and func will
return error. But in this case while handle is open impossible to do
any actions. For example call eina_file_ls for empty folder, func will
return error and fold folder open. And if we try to remove this folder
Windows only mark it to delete, and remove it after the process is
complete.

Solution: close handler in error case.

Change-Id: I178d99e2a29ef8021b35ec1e17068021363ac1f4
Signed-off-by: Vyacheslav Reutskiy <v.reutskiy@samsung.com>