Fix generate ecore key events on windows.
If Ctrl+number combination pressed/unpressed function
_ecore_win32_event_keystroke_get return NULL. It happens because
ToUnicode WinAPI func fails to prepare unicode string for given scancode
and the keyboard state.
This commit add exception for  the case with digits keys. In case if
there no translate
string, but the digit key processed - will be created a normal event as
usual.

Change-Id: Ia99feffdbc2263cc7922b9e79d836704209cfbd8
Signed-off-by: Mykyta Biliavskyi <m.biliavskyi@samsung.com>