evas: Try to fix masks of masks inside proxies
Yeah it's another fun bit of evas_render.

Issue #1:
Here's the situation:

1. A container (genlist) has a mask, M0.
2. An item I0 inside this container uses a proxy P0 as render object
   rather than the item directly (eg. for zooming in/out).
3. An element E0 inside this item has another mask, M1.

Theory:
1. The proxy surface for P0 is rendered, and M1 is applied to
   the element E0.
2. The proxy P0 is rendered on the canvas, with M0 applied.

Practice:
1. The element E0 is prepared for rendering, this triggers
   a mask subrender for M1.
2. M1 is rendered with M0 as a prev mask, then kept in cache and
   not redrawn (no geometry change, etc...)
3. When P0's surface is rendered, M1's surface is the result of M1+M0.
4. When P0 is drawn on screen, we can see the effect of M1+M0 as
   P0's geometry might be different from the item's I0.

Solution:
Discard prev masks and force a mask redraw when we're inside a
proxy. Ideally we should detect if the prev mask belongs to the
insides of the proxy or not.

Issue #2:
Somehow evas_render_mapped() messed up the rendering of a simple
mask in some rare situation. As a workaround, we can directly call
the object render function (masks are image objects anyway). This
issue could cause the disappearance of some text in a genlist.

Extras:
Remove use_mapped_ctx flag. Always assume it as true, always pass
a proper context to mapped_render. This simplifies the logic a bit.
Also avoid context free/new when we can just change two values.

This is a backport of (first tested on this tizen branch):
- 3e0f3822f144b254c9d541a32c0d789036453da4
- 1c21cdfdbb782d389818871d1697ece1098e85c2
- 2bed30b3984b4e73d64b99fe05d6574001173497
- 489815457b9e7baeadb0d8050dc25434c77d0fe9

Change-Id: I9ee857f0625a43d281eb945c2f354dd63cc0b9cd