Evas GL:Bind texture to correct one.
If user bind textureA and want to use it continuously, do not call glBindTexture(textureA) again.
But expect that textureA will be binding.
So EvasGL sould not change binded texture silently.
Restore texture to previous bound one after allocating new texture.
And when destroy texture, reset texture to 0 if it is current bound texture.

[Model] Tizen3.0
[BinType] AP
[Customer] OPEN

[Issue#] P161220-02830
[Request] PLM
[Occurrence Version] APL7

[Problem] Collage preview is showing as black while changing the ratio.
[Cause & Measure] EvasGL change the binded texture to 0 in make current job. So resture previous binded texture after allocating new texture.
[Checking Method] Reproduce issue.

[Team] UI System
[Developer] Minkyoung Kim
[Solution company] Samsung
[Change Type] Self Test

Change-Id: Ib14c2be0d69cc4f1c88ace42fc1fb4518571afbf