ecore_wl: Fixed window move and resize.
The window with property of the focus skip can not receive the
wl_keyboard::enter event, so there is no way to find the wl_seat
resource needed for the window move or resize request.

To resolve this problem it has changed to use the first wl_seat
resource in the Ecore_Wl_Display when requesting the window move
or resize.

This modification may cause unknown problems, so the tizen extension
protocol that doesn't require wl_seat resource should be added.

Change-Id: I26589c75945bc8005441597ba4b192eec4abc55f