interface_scrollable: update pan position once pan size is changed
Becase bar is closely related with pan and content,
we have to update bar position and size on some event
such as position or size of pan is changed or size of content is
changed.

However bar related things are not updated,
even if pan changed callback is invoked by pan related changing.

This behavior makes invalid drag value and wrong drag value makes
scroller had a bad position.

@tizen_fix

Change-Id: I7f013fd9a05a7f139548c4e05dd1dfc088af8b4e
Signed-off-by: Wonki Kim <wonki_.kim@samsung.com>