ecore_animator: improve comparision check for stopping animator.
Summary:
Time cmparision does not ensure the corrctness of posotion.
It is beacuse of double number calculation. Double number calculation
are not convertable.for example:
   1/39 = 0.02564102564;
   but 0.02564102564 * 39 = 0.99999999996 != 1;

The addition check for pos ensure the pos to be reached at its correct
position. this ensures the last call with pos == 1.0.

there is a check on pos at gengrid side for pos == 1 which resets animation running flag.. is not getting reset. There can be more such widgets where decisions are made on the value of pos.


@fix

Change-Id: I44681bde2ac4a7e3c08363f823cd9ba5296ba8f5