ecore_evas_extn: change indicator buffer update logic
Indicator updates its shape on one shared memory buffer.
Whenever screen is rotated, different sized shape should
be updated to the buffer. And, this causes a problem that
wrong sized indicator is shown for a while when the ecore_evas
size is changed. (i.e. the angle of display is changed)
So, this commit will skip some buffer updatings which are not
proper with current angle of the view.
Moreover, there are some codes to freeze and thaw the rendering
which will guarantee the correct updating of the buffer.

@tizen_fix

Change-Id: I7d3793eef2b2912786693fa60003e2074fb0606c