[evasgl] Use gl call immediately related frambuffer without checking
A non-NULL return value for eglGetProcAddress() does not guarante
that a function is actually supported at runtime.
so if eglGetProcAddress() return NULL, it can be problem.
this issue is especially occuring in framebuffer call.

as a matter of fact, framebuffer is core api from gles 2.0,
so it can be used immediately without checking

Change-Id: Iaec4d0f24e1c9c0d947c593688d2c09691fbf2c2