Eina hash fix backport from efl 1.20.3
We detected a problem in eina hash with ASan on x86_64.
It leads to build error in sanitized environment, but could
lead to much more problems.
It is already fixed in current EFL git:

From: "Carsten Haitzler (Rasterman)" <raster@rasterman.com>
Date: Thu, 26 May 2016 09:28:23 +0900
Subject: [PATCH] eina hash - fix stringshare key comparison function

this fixes T3638

@fix

a note... thanks so much to aerodynamik for spotting this. i'm rather
surprised coverity didn't spot this... unless someone said to "shut up
coverity you're wrong" and they should not have.

i also might have expected compilers to spot this too... and add a
warning.

anyway ... this was a seriously subtle bug that could have caused all
kinds of havoc in efl.  keys that are different may be compared to be
the same. it could get ordering wrong and sorting thus maybe insert
keys that cannot be found anymore and oh so much more besides.

Change-Id: Id81a74c1158a5b04afa26497e4033352f752689a
Signed-off-by: Dmitriy Evgenevich Gonzha <d.gonzha@samsung.com>