Fix ASan build with edje_cc on aarch64
On EFL theme build we have too many embryo_cc processes.
With ASan environment It leads to build error because of
OOM error.
Such case with unlimited embryo_cc is already fixed in
current EFL git:

From: "Carsten Haitzler (Rasterman)" <raster@rasterman.com>
Date: Mon, 13 Jun 2016 14:11:20 +0900
Subject: [PATCH] edje_cc - limit embryo_cc instances in bg to something
 sensible

so compiling current efl theme seems to spawn 178 instances of
embryo_cc in the background (then collect results). this is a bit
excessive and may be a source of issues. this may be causing T3874
perhaps due to embryo_cc's not launching (fork failing or something
else) and we don't check if fork fails. so try limit to 8 embryo_cc's
in the bg which is more reasonable than the currently "unlimited"
which leads to 178 currently...

i hope this fixes T3874

@fix

Change-Id: Ib799ca82719458b613105bbc43214b038f404827
Signed-off-by: Dmitriy Evgenevich Gonzha <d.gonzha@samsung.com>