[EvasGL] Change from function level to API in eng_window_xxx
Chanage from function level to API in eng_window_xxx.
Because some applications work with changing flag of EVAS_GL_RENDER_THREAD in runtime.
If the flag is changed in runtime, it has the effect of some global variable  as a tls...
To fix it, changing from function level to API in eng_window_xxx.

Change-Id: I27c717809fa6a1b0de100b602d73f70c635b1d4e