ecore_wl2: Modify API ecore_wl_window_geometry_get() and use it on event handler of configure.
The geometry of ecore_wl_window can be reconfigured by both of server and client.
Before, we passed its geometry as a Ecore Event in order to notify ecore_evas,
But in this case we cannot gaurantee coherence.
Thus, it would be better to give the geometry by the API on event handler.

Change-Id: Ic5dda59ab126912e0de4e4cf2d2f6c29b38f667e