ecore_evas: Pend the rotation until app set the rotation.
Currently, camera use the elm_win_rotation_with_resize_set.
it rotate the canvas, and chage the win size also.
The reason why they use this api is to use the object rotation effect during the rotation.

If window server send rotation event during the rotation, or efl client deal with the rotation
during the app's rotation effect, it cause flickering.

Before, X backend , landscape rotation
1. set the available rotation set only 0
2. window server didn't send rotation effect
3. apps listen the devicd rotation callback
4. apps set elm_win_rotation_with_resize_set 90 or 270
5. efl client set the X property related with rotation
6. window server deals with the rotation by app side.

But now window server don't want to support this api , becuase client rotation causes the whole of rotation policy.
Opensource side, server need to support this situation, but we need to time to discuss.
to support compatibility, add this code until we find the final solution.

this concept is
1. app set pending rotation using aux_hint (it means app will deal with the rotation)
2. efl client doesn't deal with server's rotation
3. do the rotation job by app just like the server's rotation.

Change-Id: I0d244194cf2f4adc0cf241b806bd1440f0026045