efl_ui_win: add APIs for aux_hint(auxiliary hint)
@tizen_feature
This is Tizen only feature.
The window auxiliary hint is the value that is used to decide which action should be made available to the user by the Windows Manager.
If you want to set a specific hint to your window, then you should check whether it exists in the supported auxiliary hints.

Change-Id: Ibe51b79ac74771204c4ac09b003c9ddf7cb6c64f