[evasgl] Delete checking checksum when copy old shader
when application use runtime shader, it add runtime shader when application shutdown.
for add runtime shader, they copy old shader file to temp file  and modify,rename this.
but in this case, eet_open & checksum check is failed.
because temp file is empty (it didn't write checksum)
so delete checking checksum when copy old shader to temp file.

and..

if new shader is compiled, it will be saved after shutdown.
because of this, shader is compiled 2~3 times in some case.
so add some code for this problem.
1. close & reopen shader before checking shader file.
2. update shader file when application compile new shader

Change-Id: I3eb966d9f67695cb576db6b955824574251083b6