atspi: add state_notify api
The purpose of this api is to provide additional way of synchronization
between elementary application and assistive technology client
(screen reader).

The subject of synchronization is accssibility object state which
is required by screen reader implementation to work properly.

Because of nature of efl, visibility changes of efl object state cannot be
effectively evaluated in a moment they occur due to following obstacles:

1. requires tracking of visibility of parent, parent's parent etc.
2. requires tracking of clipper of parent, parent's parent etc.
3. requires tracking of all clippers movement
4. requires tracking of all clippers alpha
5. requires tracking of stacking order 5. requires tracking of object movement.

 Such implementation will have heavy impact on framework.
To cover more cases an additioanl method of synchronization is provided
which can notify assistive technology clients about state of accessible
objects at the moment it is invoked.

This is meant to be elementary internal api.

@tizen_feature

orignal patch : 10aa19137bb1c6168e26f0973aa06caab6942fbf

Change-Id: Icc41b643b72fb2fb0ccd438394f71749de533fa5