elm: [atspi]check if highlighted object's parent hidden
If an object is content of a layout (parent), and user calls
evas_object_hide for the layout, then the _elm_widget_evas_object_smart_hide
for the object (child) is not called. So it is not possible to know
object's changed state on the AT-client side.

 This patch set will check parents of highlighted object in
_elm_widget_evas_object_smart_hide, and emit state changed
signal if one of parents is equal to the hiding object.

@tizen_fix

orignal patch: afcec90361769e19f6d6dc005aa35daeaa74b8ec

Change-Id: I45d7d71c96e1a2645c93f2d1012067f880d6a99d