elm: access: emit signal after atspi bridge is connected
This patch set emitting popup SHOWING signal, and window ACTIVE signal after
the atspi bridge is connectted.

The SHOWING signal of popup does not work properly, if the popup shows before
the atspi bridge is not connected. In this case, the screen-reader cannot read
popup information. So the SHOWING signal should be emitted after the connection
is established.

The ACTIVE signal does not work properly in the following case.
There are two windows show in one process at the same time. In this case,
the ACTIVE signal was not emitted because window_activated_broadcast_needed is
set to EINA_FALSE. This would be a bug.

orig: fe8b20fac57caa39c0f801d3819e688f391c1831

Change-Id: I7e88b20c1d31ecc41410d8d81159654ce637e133
Signed-off-by: Shilpa Singh <shilpa.singh@samsung.com>