Elm focus navigation: other 2 failing tests
Added a new test "Focus 6", it's an not very
complex elm layout: a swallowed genlist and three
buttons in an edje box.

You should be able to navigate the layout with
just the keyboard, that is currently impossible.

With the help of the mouse click you can randomly
make the key navigation work again... this is
mostly random.

...should help to make progress on T6453