config: add "language_auto_mirrored"
when language_auto_mirrored is true, elm_config_mirrored_set() will be
called automatically according to "default:LTR" value in elm_language_set() function.
for initialization, it needs to be done in translation_init() as well,
but according to appcore issue, we do it in elm_win_add() time temporarily.
when it is false, mirroring will be done manually by calling
elm_config_mirrored_set().
this patch will take effect with efl-misc package.

@tizen_feature

Change-Id: I28524dbc1f2c6286d0d1a1e5f3e7af729dd0068e