ecore_con + extn: request E compositor to create socket
In Tizen, the security does not allow 1 application to create
socket and other applications to connect to that socket.
All sockets should be created by system daemon or service.
Therefore, we request the E compositor to create sockets for us.
Besides, this patches also change the location of sockets to
avoid smack errors.

ref 4d81d8f1ef043853f05494005ab4c46ffe2a4599 and
    c173e3d4e4dcea93f7589b9360ea99d448e08301
@tizen_feature

Change-Id: I584c51632053e8e9f576ec69b8c4026f930c2be6