evas font: Don't pick a script font based on COMMON/UNKNOWN script text
When there was a COMMON/UNKNOWN script text, Evas picked a script font only for that text.
And when a explicit character was appened to the text, all previous text before new character
would loaded from new script font according to the new character.

Issue Example)
1. [EMOJI] + [SPACE]
   All script is COMMON/UNKNOWN. A script font is selected as "BreezeColorEmoji" font.
   All text will be loaded from "BreezeColorEmoji" font which does not have proper space character.
2. [EMOJI] + [SPACE] + [LATIN "A"]
   A script font is selected as "BreezeSans" font because of LATIN "A" character.
   The emoji characters will be loaded from "BreezeColorEmoji" font.
   But, the space character and LATIN "A" character will be loaded from "BreezeSans" font.

So, this patch fixes to pick a top font as its script font when the given script is not explicit script.

@tizen_fix

Change-Id: I31aeb80f8e8828bb8709d549b7178c83cfcbcd7c