elm: do not set input region
If a window is not a full screen, then the window has used
geometry of "elm.spacer.opaque" part for its input region.

This geometry value is not correct at first launching time,
if window does not have its resize object.

For example the keyboard shows with "0, 0, 63, 0" input region,
and the keyboard could not get touch events from the window manager.

*tizen_only

Change-Id: I1824649f67674befe59514fa119fb5c000013580