atspi: use correct object for highlight_grabbing
elm_object_accessibility_highlight_set is using VIEW object as a "target"
object for the ELM_WIDGET_ITEM_CLASS object. And the "target" object is used
for all highlight related lines in the elm_win including highlight_grabbing.

So when the elm_widget sets highlight_grabbing for the ELM_WIDGET_ITEM_CLASS
object, the VIEW object should be used.

Change-Id: Id7f5595193a140e6b3d8f2c9c5a94ed9a505d890