atspi: not accept redundant leaf object
An object which role is ELM_ATSPI_ROLE_REDUNDANT_OBJECT, has been acceptable
so far regardless its children. So it was possible to find its children.
But it does not make sense to accept ELM_ATSPI_ROLE_REDUNDANT_OBJECT object if
it does not have children.

There was an issue on Browser application because we accepted REDUNDANT leaf
object. There is a button which role is ELM_ATSPI_ROLE_REDUNDANT_OBJECT, this
button would be used for conveying focus event to embedded web page.
The button was accepted, so it was not possible grab an object on web page.

Please refer to following part of accessibility tree. It is not possible to
get Elm_Atspi_Ewk_Wrapper, if Elm_Button is accepted for "GetNavigableAtPoint"

Elm_Layout
     - Elm_Button [redundant object][0,148,1080,1604]
     - Elm_Atspi_Ewk_Wrapper [embedded][0,148,1080,1604]

Change-Id: I382ececdce1e0e849d9040d6a5f2ba29cfaccb99