travis: fix install of packages on macOS
Some recent OSX builds failed on Travis due to a change on the macOS
image they are providing. The python package could not be installed.

==> Pouring python@2-2.7.14_3.sierra.bottle.1.tar.gz
Error: The `brew link` step did not complete successfully
The formula built, but is not symlinked into /usr/local
Could not symlink bin/2to3-2
Target /usr/local/bin/2to3-2
is a symlink belonging to python. You can unlink it:
  brew unlink python