elementary layout: just return false when the given pos is negative
The function for this feature in Edje starts supporting negative pos.
The negative pos will be used for inserting the given object at backward position.

But, in Tizen 4.0, the document for this function described pos has to be positive or zero.
Actually, it failed when pos was negative.
So, for backward compatibility, it must return FALSE in this case.

@tizen_fix

Change-Id: I004af7839976a94b5e4e68a64c795689f0a25ca3