Genlist: visual test for item next/prev/first/last
Added 4 buttons to the "genlist group tree" to perform
item next/prev/first/last.

This test works correctly by reverting fd82c2521ebb but it
still in contrast with the item_next() docs, that say item_next
should not descend in item childs.

I think that the docs is also wrong, as I never saw the behaviour
that is explained there.