ecore_evas: use timer for animator
[Problem]
The animator on the wayland ecore_evas engine used wl_surface frame callback
as its source(of _ecore_evas_tick_source_find). This made wl_surface_commit
be called too much and caused the Enlightenment performance issue.

[Solution]
So this patch set is using the timer for animator source.
The animator is working without intervention of the Enlightenment.

@tizen_only

Change-Id: Id20a95a35107454a908397eeb3640d7f6650ef2d