eo: get rid of eo_id
Summary:
build with this disabled is just not working as the thread model of eo
falls apart. The threadmodel is required to have eo_id, as it decoded
the thread information in the eo_id, which is not working without eo_id.

This also fixes the testcases that have never been executed due to the
fact of the missing HAVE_EO_ID

fix T6610

Depends on D6327

Reviewers: devilhorns, zmike

Reviewed By: zmike

Subscribers: cedric, #committers, zmike

Tags: #efl

Maniphest Tasks: T6610

Differential Revision: https://phab.enlightenment.org/D6328