evas_drm: Allocate buffers on demand
Summary:
Instead of allocating a fixed number of buffers immediately, allocate
buffers if needed to render to.

Normally we only need 2 buffers, but we've been allocating 3 to handle
worse case behaviour.  As T7099 shows, this is not always enough.  We
now cap at a max of 10.

For the normal case where we always use 2 this results in a slight
memory reduction (1 buffer) and a slight renering load reduction
because we pick the oldest buffer to render into.

A future patch will trim the buffer queue if it's been too large for
a long time.

fix T7099
Depends on D6563

Reviewers: devilhorns

Reviewed By: devilhorns

Subscribers: cedric, #committers, zmike

Tags: #efl

Maniphest Tasks: T7099

Differential Revision: https://phab.enlightenment.org/D6564