ecore: fix order of signal fd shutdown
Summary:
When closing the fd handler, it checks if the fd is already closed and prints
an annoying  warning: "fd %d closed, can't remove from epoll - reinit!"
We need to close the handler first and then the actual fd.
I am not familiar with this part of the code, but this fix removes the warnings
and does not seems to have adverse effects.

Test Plan: It had warnings before and now it doesn't, haven't observed any other adverse effect.

Reviewers: raster, zmike, devilhorns

Reviewed By: zmike

Subscribers: cedric, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D6561