elua: make old documentation generator generate instead of failing
While a rewritten template-based doc generator is pending addition
into a separate repository, at least fix this one so that it works
in case we decide to go back to it, because right now it's just
failing because of out of date code.

The fixes here are based on the revamped documentation generator
that will be introduced.