ecore_con_local: change location for socket.
In Tizen, the security does not allow 1 application to create
socket and other applications to connect to that socket.
All sockets should be created by system daemon or service.
Therefore, we request the E compositor to create sockets for us.
Besides, this patches also change the location of sockets to
avoid smack errors.

@tizen_only

Change-Id: I9ceda7bbd5cb7b06331bb2ced6342f646a2f44fb