genlist: fix item show/bring in coordinate calculation with zero-sized pan.
item show and bring can be processed with zero-sized pan.
item's coordinate will be calculated with this zero-sized pan,
so target position of scroll region bring in is not proper.

it occurs wrong result of SCROLLTO_MIDDLE and SCROLLTO_BOTTOM cases.

now we check pan size, and if less than 1, deferred call.

this is port of upstream patch D7037

Change-Id: Ie5d5ecd039d835caefa47c272547506946e91a37
Signed-off-by: SangHyeon Jade Lee <sh10233.lee@samsung.com>