Meson: link elementary_test with rdynamic
Summary:
This is needed to let dlsym extract the efl_main/main... symbols. For
example, the Exactness player needs it to launch a given binary. If
this option is not provided during the linkage of the binary, the
player won't be able to run it.

Test Plan: exactness_play -t <path to exactness scenarios>/radio.rec -- elementary_test -to radio

Reviewers: bu5hm4n

Reviewed By: bu5hm4n

Subscribers: cedric, #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D7178