ecore event_loop: add the error handler for event_loop
When the return value of select is ZERO, _ecore_main_fdh_mark_active did not called.
So in _ecore_main_awake_handler, we cannot guarantee right operation. For this problem,
Even the return valie of select is ZERO, call the _ecore_main_fdh_mark_active()
with NULL parameters. and In _ecore_main_fdh_mark_active(), reset the all flags for each
fd_handlers.

Change-Id: I6f47f6ce9f824a21e668c099d1601243b3af8b53