travis: move parts of our build matrix to only run on the daily travis cron job
We need to keep our builds running for every push to a minimum. Various
distro builds as well as the release-ready build can happily run once a
day.
This commit also switches from a build matrix to a simple list of build
jobs to allow the usage of build type = cron condition (not possible
with the matrix builds)

Differential Revision: https://phab.enlightenment.org/D7293