build: ensure we have a data_aux direcroy in our dist for the test eo files
Summary:
Due to our EXTRA_DIST2 crutch we need to handle the directory creation
on our won here. That was forgotten for the new data_aux folder in the
eolian tests. Creating it allows the files to be copied over and the
eolain seuite to pass in distcheck.

Reviewers: q66

Reviewed By: q66

Subscribers: cedric, #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D7477