efl_ui_widget: refactor sub_object handling
efl_ui_widget has a property called widget_parent. The setter for this
function is called is exactly once, and this is within the constructor,
to a value which is not even set to the actaul field parent_obj. Which
shows, that in the sitation right now, the setter of the property is a
bit disconnected and lags some real aspects.
As we are heading towards eo-api stabilization we should beat some sense
into this setter, as people using our classes might overwrite the setter
and except calls to it, whenever the widget_parent is changed, and
implementation as in elm_menu show that this might makes sense sometime.

In order to achive this, the sub_object registering code of elm is
adjusted a bit.
sub_object_add/del is now used to differenciate between evas objects and
efl.ui.widget objects as subobject. In case of a widget, the
widget_parent of this object is set, most of the widget specific code is
then executed in the actaul setter. In case of an evas object, the
parent reference is added. In the end both end up in the subobject
children list. The later is also a requirement for widget_parent_set to
be successfull.

ref T7553

Reviewed-by: Mike Blumenkrantz <michael.blumenkrantz@gmail.com>
Differential Revision: https://phab.enlightenment.org/D8031

Change-Id: I32ee38e867668e8fef486dc4807d42cc2c3a589c