efl_ui_widget: remove widget_parent overwritting
before the widget_parent have been only set once. The call to the set
was in the constructor and carried the efl_parent. In the next commit
widget_parent is getting a refactor, which gives it more meaning, where
it is actaully called, which means, the behaviour will change. In order
to not break every existing usage of the here changed widgets, we move
the code to the constructor, and feed it with the efl_parent, just like
before.

Differential Revision: https://phab.enlightenment.org/D8041

Change-Id: Id39fb73831c45d33b6c419316e806fafec9cbfaf