meson - fix some more fbsd building and make notes on broken checks
i found the header checks are logically broken. i was about to replace
with a cc.compiles() check but that requires that every check also
have full working code not just a symbol to check the existence form
so for now just fix the HAVE_NOTIFY_KEVENT define that makes eio build
properly again on bsd and note the broken state for a future rethink
on how these checks are done.