elementary: Add ELM_WIDGET_DATA_GET macro
There are cases, such as 2abe5ad9bc6d30f0c29381dca7e where we do not
always want to return if we fail to get the elm widget data (as we may
have to cleanup leaks, etc). As such, add a macro we can call which
will just try to fetch the elm_widget data.